+++ memory-viz/src/memory_model.ts
@@ -91,7 +91,7 @@ export class MemoryModel {
         this.roughjs_config = options.roughjs_config ?? {};
         this.rough_svg = rough.svg(this.svg, this.roughjs_config);
 
-        setStyleSheet(this);
+        setStyleSheet(this, options.global_style ?? "");
 
         // The user must not directly use this constructor; their only interaction should be with 'user_functions.draw'.
         for (const key in config) {