+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### ✨ Enhancements
 
+- Added support for global CSS styling using the `--global-style` option in the CLI
+
 ### 🐛 Bug fixes
 
 - Fixed bug where the box wouldn't render properly when value is null for primitive types